# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ZohoConnectionStatus {
  connected: boolean;
  /** @nullable */
  organizationId?: string | null;
  /** @nullable */
  organizationName?: string | null;
}

export interface ZohoInvoice {
  invoiceId: string;
  invoiceNumber: string;
  customerName: string;
  /** @nullable */
  description?: string | null;
  status: string;
  invoiceDate: string;
  /** @nullable */
  dueDate?: string | null;
  total: string;
  balance: string;
  currencyCode: string;
  syncedAt: string;
}

export interface ZohoSyncResult {
  syncedCount: number;
  syncedAt: string;
}

export type DemoBankSelectionInstitutionId = typeof DemoBankSelectionInstitutionId[keyof typeof DemoBankSelectionInstitutionId];


export const DemoBankSelectionInstitutionId = {
  hdfc: 'hdfc',
  icici: 'icici',
} as const;

export interface DemoBankSelection {
  institutionId: DemoBankSelectionInstitutionId;
}

export interface BankAccount {
  accountId: string;
  provider: string;
  institutionId: string;
  institutionName: string;
  accountType: string;
  last4: string;
  currentBalance: string;
  currencyCode: string;
  status: string;
  lastSyncedAt: string;
  importCompleted: boolean;
  importedTransactionCount: number;
}

export interface BankConnectionStatus {
  connected: boolean;
  onboardingComplete: boolean;
  importedTransactionCount: number;
  accounts: BankAccount[];
}

export interface BankImportResult {
  importedCount: number;
  importedAt: string;
}

export interface BankTransaction {
  transactionId: string;
  accountId: string;
  transactionDate: string;
  description: string;
  debit: string;
  credit: string;
  balance: string;
  reference: string;
  currencyCode: string;
}