# lib/api-zod/src/generated/types/bankAccount.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BankAccount {
  accountId: string;
  provider: string;
  institutionId: string;
  institutionName: string;
  accountType: string;
  last4: string;
  currentBalance: string;
  currencyCode: string;
  status: string;
  lastSyncedAt: Date;
  importCompleted: boolean;
  importedTransactionCount: number;
}