# lib/api-zod/src/generated/types/bankConnectionStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BankAccount } from './bankAccount';

export interface BankConnectionStatus {
  connected: boolean;
  onboardingComplete: boolean;
  importedTransactionCount: number;
  accounts: BankAccount[];
}