# lib/api-zod/src/generated/types/bankImportResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BankImportResult {
  importedCount: number;
  importedAt: Date;
}