# lib/api-zod/src/generated/types/bankTransaction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BankTransaction {
  transactionId: string;
  accountId: string;
  transactionDate: string;
  description: string;
  debit: string;
  credit: string;
  balance: string;
  reference: string;
  currencyCode: string;
}