# lib/api-zod/src/generated/types/demoBankSelection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DemoBankSelectionInstitutionId } from './demoBankSelectionInstitutionId';

export interface DemoBankSelection {
  institutionId: DemoBankSelectionInstitutionId;
}